backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
httpx
anyio
pydantic
numpy
scipy
sympy
